Criadas as regras de negócio da tabela produtos

+++ src/services/deleteProductService.ts
@@ -5,6 +5,12 @@ class DeleteProductService {
   async execute(id: string) {
     const productRepository = getCustomRepository(ProductsRepository);
 
+    const productExists = await productRepository.findOne({ id });
+    
+    if(!productExists) {
+      throw new Error("Produto não encontrado, operação não realizada");
+    }
+
     await productRepository.delete({ id });
 
     return;
